Bug fix: Let servisor honour the number of workers to fork
A tiny naming bug prevented servisor from starting the sought number of
workers. This commit fixes it. Concretely, if the user provides that
number as a command-line argument, it is honoured. Otherwise, the
num_workers stanza from the configuration file is used. Furthermore,
clients can now set num_workers to 'ncpu' in the config, which causes
servisor to start as many workers as there are available CPU units on
the machine.