Resolve paths relative to the module location
In production services are generally started with node /full/path/to/server.js
without changing the working directory to the actual service checkout. This
means that resolving module include paths relative to the cwd won't work.

This patch changes the resolution to one relative to the service-runner
binary. If it is located inside node_modules/service-runner, that the
directory containing node_modules is used. Otherwise, the current working
directory is used.