Keep reporting the stats even when the heap limit has been exceeded
When a worker exceeds its heap limit, it tries to disconnect, otherwise
it is forcibly removed. In either case, no more statistics are being
reported, so we cannot know exactly when it has exited and how did the
heap size change in the last moments. This commit allows workers to keep
reporting their stats until they die even if they have been designated
for exclusion.