Ensure different streams log levels will be respected
The logger uses the _levelMatcher private property to determine if a log
entry should be logged or not. However, the threshold is set based on
conf.level, without taking into account individual stream levels. That
causes log entries for levels lower than the global level not to be
logged in the individual log streams.

This commit performs the check to find the minimum log level to use
correctly based on individual stream settings. Additionally, it
explicitly appends the level property to streams lacking it.