Don't discard most logging info in logger
bff5450 changed the logic to only log a wrapped error instance & drop all
other information on the floor. This is problematic, as this means that
details like internal requests don't make it into the logs, greatly
complicating debugging.

This patch restores the previous behavior of logging all provided information,
but also makes sure that the msg parameter is provided as a positional
parameter to bunyan.