Docker script: De-duplicate module dependencies
Most packages use ^ and/or ~ to designate suitable dependencies.
Coupling that with npm's way of searching for packages recursively up
the path means that there is great benefit in de-duplicating all of
service's packages when running in the container. Most notably, this
vastly simplifies and minimises the deploy repository's footprint.