Add the ability to create prefixed child metric clients
We would like to be able to prefix all metrics emitted as part of a request
with a fixed prefix, so that we can distinguish internal / external or update
requests & all sub-requests associated with it.

Similar to the logger instance, this patch adds a `makeChild` method on
returned metric clients, which given a name will create such a prefixed metric
client.