Release service-runner v1.0.0
We have been using service-runner in production to run at least half a dozen
services for many months, and it has proven very reliable. The interfaces are
stabilized as well, so it's time to commit to version one-dot-zero.

This also applies regular semver policy to our version increments, rather than
relying on npm's special case behavior for major < 0.

Also, there is a small bug-fix for child metric prefixes.