Fix service module loading and release v1.1.4
One can specify an npm module name to load for a service's start-up
process. Node then tries to find it in the usual way - by searching the
path up to root for that module and in the global modules' directory.
However, node starts the search from service-runner's path, not the
service's one. Effectively, the module that service-runner is requiring
is its peer dependency. Unfortunately, this doesn't work in cases where
node modules are symlinked into the relevant node_modules directory. So,
when a relative path is encountered and the module fails to load, try to
load it both from the application's base path as well as its
node_modules sub-directory, if one exists.