Bail out of module loading if the error is not about the module (#100)
The `_requireModule()` method tries all possible combinations of a
module's path in order to find it and load it. However, up until now we
would try to do so regardless of the reason of module failure, which
meant that if the module could be found, but there was an error while
loading it, we would assume it couldn't be found. This resulted in
cryptic error messages of the type "Cannot find module
$PWD/node_modules/module.js" even when the module could be found, but
failed to load for some other reason. This commit fixes it.