Docker: Allow only the Dockerfile to be generated
As a first step of the containerasition effort, we need to make the
Dockerfile specification based on package.json available to other
parts of the system, most notably the build and test pipeline(s). To
that end, service-runner needs to be able to only generate the spec
without actually building nor starting the container.

This commit introduces the `generate` command which allows one to only
create the Dockerfile specification.