Docker: Allow explicit base image names in pkg.deploy.target
So far, the exact target system was looked for through a combination of
the target specified in the dpeloy section and the list of available
targets. However, that can get tricky if the service doesn't have the
targets.yaml file and when there are multiple versions of the same
distribution. Hence, allow the user to specify the exact target to use.
It is recognised by looking for ":" in the target name. This commit also
fixes a bug whereby the target image was being deduced before the
defaults were set.

Additionally, do not install npm by default, but rather build-essential.
The latter is always needed, whereas the former is needed only in cases
where the system nodejs package is used. Also, update nvm to v0.33.6

Finally, update the linting dependencies and fix some linting errors.