Allow user to specify custom npm version
Sometimes npm shipped with node has bugs that break the build
for certain packages. To be able to quickly workaround such bugs
we allow to select a custom npm version that should be used.

Also, the reshrinkwrap setting was never used since it's creation
and is not really needed in the future with introduction of
packages-lock.json file in npm v4, so it was removed.