Fix Docker working directory permissions
The current Dockerfile leaves root as the owner of the working directory,
which can cause permissions issues in the event that the run user created
on non-Mac platforms needs to modify files in the working directory.
Running docker-test for mobileapps on Linux currently fails for this
reason.  This patch adds a command to change ownership of the working
directory to runuser:rungroup.