add staticLabels feature
the convention in k8s-converted applications is to have a "service" label associated with
the more system-level metrics like heap, gc, and request duration (like in hyperswitch).

this feature enables the user to define static labels for Prometheus metrics and
exposes a new getter on the Metrics class to get the service name it knows about.