Add support for having named logging levels
Why:
service-runner, when logging to stdout delegates to bunyan stdout
logging process that ends up creating log entries with level specified
as an integer, even though the code everywhere might be using named
levels. This hasn't been a problem up to now because we sparringly
relied on that, but with the move to service-runner apps on kubernetes
and the logging pipeline this is starting to cause issues, namely
https://phabricator.wikimedia.org/T239459. It would be nice if
service-runner allowed to log to stdout named loglevels

What:
Add a simple, albeit probably inefficient function due to all the JSON
encoding/decoding that wraps process.stdout and changes the level
attribute from an integer to a string. Provide it as an additional
stream.

The output looks like this:

{"name":"myapp","hostname":"buster","pid":5835,"level":"INFO","msg":"This is logging","time":"2020-04-08T09:03:09.294Z","v":0}

instead of:

{"name":"myapp","hostname":"buster","pid":5835,"level":30,"msg":"This is logging","time":"2020-04-08T09:03:09.294Z","v":0}

Bump service-runner version to reflect the change

Bug: T239459